refactor: consolidate DropdownItemType enum and clean up imports across components
